# openapi/api-v1.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "RPP AI Mobile API",
    "version": "1.0.0-target",
    "summary": "Kontrak target API v1 untuk Laravel dan Android",
    "description": "Source of truth implementasi mendatang. Dokumentasi ini tidak menyatakan endpoint telah tersedia pada runtime saat ini."
  },
  "servers": [
    {
      "url": "{baseUrl}/api/v1",
      "variables": {
        "baseUrl": {
          "default": "http://localhost:8000"
        }
      }
    }
  ],
  "security": [
    {
      "sanctum": []
    }
  ],
...
  "paths": {
...
    "/rpps": {
      "get": {
        "tags": [
          "RPP"
        ],
        "operationId": "listRpps",
        "summary": "Daftar RPP",
        "responses": {
          "200": {
            "description": "Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Rpp"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          },
          "422": {
            "$ref": "#/components/responses/E422"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          }
        ]
      },
      "post": {
        "tags": [
          "RPP"
        ],
        "operationId": "createRpp",
        "summary": "Buat RPP",
        "responses": {
          "202": {
            "description": "Diterima untuk diproses",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Rpp"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          },
          "404": {
            "$ref": "#/components/responses/E404"
          },
          "422": {
            "$ref": "#/components/responses/E422"
          },
          "429": {
            "$ref": "#/components/responses/E429"
          },
          "500": {
            "$ref": "#/components/responses/E500"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RppCreate"
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ]
      }
    },
    "/rpps/{id}": {
      "get": {
        "tags": [
          "RPP"
        ],
        "operationId": "getRpp",
        "summary": "Detail RPP",
        "responses": {
          "200": {
            "description": "Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Rpp"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          },
          "404": {
            "$ref": "#/components/responses/E404"
          },
          "422": {
            "$ref": "#/components/responses/E422"
          },
          "429": {
            "$ref": "#/components/responses/E429"
          },
          "500": {
            "$ref": "#/components/responses/E500"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ]
      },
      "delete": {
        "tags": [
          "RPP"
        ],
        "operationId": "deleteRpp",
        "summary": "Hapus RPP",
        "responses": {
          "204": {
            "description": "Dihapus"
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          },
          "404": {
            "$ref": "#/components/responses/E404"
          },
          "422": {
            "$ref": "#/components/responses/E422"
          },
          "429": {
            "$ref": "#/components/responses/E429"
          },
          "500": {
            "$ref": "#/components/responses/E500"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ]
      }
    },
...
    "/settings": {
      "get": {
        "tags": [
          "Settings"
        ],
        "operationId": "getSettings",
        "summary": "Pengaturan sekolah",
        "responses": {
          "200": {
            "description": "Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Settings"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          },
          "404": {
            "$ref": "#/components/responses/E404"
          },
          "422": {
            "$ref": "#/components/responses/E422"
          },
          "429": {
            "$ref": "#/components/responses/E429"
          },
          "500": {
            "$ref": "#/components/responses/E500"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Settings"
        ],
        "operationId": "updateSettings",
        "summary": "Perbarui pengaturan sekolah (capability settings.manage)",
        "responses": {
          "200": {
            "description": "Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Settings"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          },
          "404": {
            "$ref": "#/components/responses/E404"
          },
          "422": {
            "$ref": "#/components/responses/E422"
          },
          "429": {
            "$ref": "#/components/responses/E429"
          },
          "500": {
            "$ref": "#/components/responses/E500"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/SettingsInput"
              }
            }
          }
        }
      }
    },
...
    "/admin/rpps": {
      "get": {
        "tags": [
          "Admin RPP"
        ],
        "operationId": "listAdminRpps",
        "summary": "Daftar RPP guru",
        "responses": {
          "200": {
            "description": "Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Rpp"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          },
          "404": {
            "$ref": "#/components/responses/E404"
          },
          "422": {
            "$ref": "#/components/responses/E422"
          },
          "429": {
            "$ref": "#/components/responses/E429"
          },
          "500": {
            "$ref": "#/components/responses/E500"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "sanctum": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Sanctum"
      }
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 16,
          "maxLength": 128
        },
        "description": "Kunci unik per intent; replay identik mengembalikan resource sama, payload berbeda menghasilkan 409."
      }
    },
    "responses": {
      "E401": {
        "description": "Tidak terautentikasi",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "E403": {
        "description": "Tidak diizinkan",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "E404": {
        "description": "Tidak ditemukan",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "E422": {
        "description": "Validasi gagal",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "E429": {
        "description": "Terlalu banyak permintaan",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "E500": {
        "description": "Kesalahan server",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {},
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "links": {
            "$ref": "#/components/schemas/PaginationLinks"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "message",
          "code"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "current_page": {
            "type": "integer"
          },
          "last_page": {
            "type": "integer"
          },
          "per_page": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "from": {
            "type": "integer"
          },
          "to": {
            "type": "integer"
          }
        }
      },
      "PaginationLinks": {
        "type": "object",
        "properties": {
          "first": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "last": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "prev": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "next": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          }
        }
      },
...
      "AsyncStatus": {
        "type": "string",
        "enum": [
          "processing",
          "completed",
          "failed"
        ]
      },
      "Rpp": {
        "type": "object",
        "required": [
          "id",
          "user_id",
          "nama_guru",
          "mata_pelajaran",
          "fase",
          "topik",
          "status"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "nama_guru": {
            "type": "string"
          },
          "mata_pelajaran": {
            "type": "string"
          },
          "fase": {
            "type": "string"
          },
          "kelas": {
            "type": [
              "string",
              "null"
            ]
          },
          "semester": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "Ganjil",
              "Genap",
              null
            ]
          },
          "topik": {
            "type": "string"
          },
          "kurikulum": {
            "type": "string"
          },
          "tema": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/AsyncStatus"
          },
          "content_result": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": true
          },
          "failure_message": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RppCreate": {
        "type": "object",
        "required": [
          "nama_guru",
          "mata_pelajaran",
          "fase",
          "topik",
          "alokasi_waktu",
          "kurikulum"
        ],
        "properties": {
          "nama_guru": {
            "type": "string",
            "maxLength": 255
          },
          "kepala_sekolah": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 255
          },
          "nip_kepala_sekolah": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 50
          },
          "kota": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 100
          },
          "tanggal": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "mata_pelajaran": {
            "type": "string",
            "maxLength": 255
          },
          "fase": {
            "type": "string",
            "enum": [
              "A",
              "B",
              "C",
              "D",
              "E",
              "F",
              "RA",
              "MI Rendah",
              "MI Tinggi",
              "MTs",
              "MA"
            ]
          },
          "kelas": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 20
          },
          "semester": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "Ganjil",
              "Genap",
              null
            ]
          },
          "target_peserta_didik": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 100
          },
          "topik": {
            "type": "string",
            "maxLength": 1000
          },
          "alokasi_waktu": {
            "type": "string",
            "maxLength": 100
          },
          "jumlah_pertemuan": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 1,
            "maximum": 10
          },
          "kompetensi_awal": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 1000
          },
          "kata_kunci": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 500
          },
          "model_pembelajaran": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 255
          },
          "jenis_asesmen": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "Diagnostik Kognitif",
                "Diagnostik Non-Kognitif",
                "Formatif",
                "Sumatif"
              ]
            }
          },
          "kurikulum": {
            "type": "string",
            "maxLength": 255
          },
          "tema": {
            "type": [
              "string",
              "null"
            ]
          },
          "panca_cinta": {
            "type": "boolean"
          },
          "adiwiyata": {
            "type": "boolean"
          },
          "kka": {
            "type": "boolean"
          }
        }
      },
...
      "Settings": {
        "type": "object",
        "properties": {
          "nama_sekolah": {
            "type": [
              "string",
              "null"
            ]
          },
          "nsm": {
            "type": [
              "string",
              "null"
            ]
          },
          "npsn": {
            "type": [
              "string",
              "null"
            ]
          },
          "alamat": {
            "type": [
              "string",
              "null"
            ]
          },
          "logo_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "logo_kanan_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          },
          "kop_surat_url": {
            "type": [
              "string",
              "null"
            ],
            "format": "uri"
          }
        }
      },
      "SettingsInput": {
        "type": "object",
        "properties": {
          "nama_sekolah": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 255
          },
          "nsm": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 20
          },
          "npsn": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 20
          },
          "alamat": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 500
          },
          "logo": {
            "type": [
              "string",
              "null"
            ],
            "format": "binary"
          },
          "logo_kanan": {
            "type": [
              "string",
              "null"
            ],
            "format": "binary"
          },
          "kop_surat": {
            "type": [
              "string",
              "null"
            ],
            "format": "binary"
          }
        }
      },
...
    }
  }
}